+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Deprecated `git_revision` function in favor of new `git_revision_auto`
+
+
 0.1.2
 -----
 - Deprecated `get_revision` function in favor of new `git_revision`